chore(suite): separate popup logic from `@suite-common/connect-init`
