+++ packages/suite-desktop-connect-popup/src/index.ts
@@ -1,0 +1,1 @@
+export * from './connectPopupThunks';
