+++ packages/suite-desktop-connect-popup/tsconfig.json
@@ -1,0 +1,23 @@
+{
+    "extends": "../../tsconfig.base.json",
+    "compilerOptions": { "outDir": "libDev" },
+    "references": [
+        {
+            "path": "../../suite-common/redux-utils"
+        },
+        {
+            "path": "../../suite-common/suite-types"
+        },
+        {
+            "path": "../../suite-common/test-utils"
+        },
+        {
+            "path": "../../suite-common/wallet-core"
+        },
+        { "path": "../connect" },
+        { "path": "../env-utils" },
+        { "path": "../suite-desktop-api" },
+        { "path": "../urls" },
+        { "path": "../utils" }
+    ]
+}
