+++ packages/suite/tsconfig.json
@@ -104,6 +104,9 @@
         { "path": "../suite-analytics" },
         { "path": "../suite-data" },
         { "path": "../suite-desktop-api" },
+        {
+            "path": "../suite-desktop-connect-popup"
+        },
         { "path": "../suite-storage" },
         { "path": "../theme" },
         { "path": "../type-utils" },
