+++ suite-common/connect-init/tsconfig.json
@@ -8,9 +8,6 @@
         { "path": "../wallet-core" },
         { "path": "../../packages/connect" },
         { "path": "../../packages/env-utils" },
-        {
-            "path": "../../packages/suite-desktop-api"
-        },
         { "path": "../../packages/urls" },
         { "path": "../../packages/utils" }
     ]
